changed BUFFER_SIZE to usize (cf. #49)

+++ src/node/delay.rs
@@ -64,8 +64,8 @@ impl DelayNode {
 
             // allocate large enough buffer to store all delayed samples
             let max_samples = options.max_delay_time * context.base().sample_rate().0 as f32;
-            let max_quanta = (max_samples.ceil() as u32 + BUFFER_SIZE - 1) / BUFFER_SIZE;
-            let delay_buffer = Vec::with_capacity(max_quanta as usize);
+            let max_quanta = (max_samples.ceil() as usize + BUFFER_SIZE - 1) / BUFFER_SIZE;
+            let delay_buffer = Vec::with_capacity(max_quanta);
 
             let render = DelayRenderer {
                 delay_time: proc,
@@ -116,7 +116,7 @@ impl AudioProcessor for DelayRenderer {
         let delay = params.get(&self.delay_time)[0];
 
         // calculate the delay in chunks of BUFFER_SIZE (todo: sub quantum delays)
-        let quanta = (delay * sample_rate.0 as f32) as usize / BUFFER_SIZE as usize;
+        let quanta = (delay * sample_rate.0 as f32) as usize / BUFFER_SIZE;
 
         if quanta == 0 {
             // when no delay is set, simply copy input to output
